Time: 2 ms (4.11%), Space: 42.3 MB (62.34%) - LeetHub

+++ 3622-check-divisibility-by-digit-sum-and-product/3622-check-divisibility-by-digit-sum-and-product.java
@@ -3,7 +3,7 @@ class Solution {
         int a = 0;
         int b = 1;
         for(char c : (n + "").toCharArray()){
-            int m = (int) c - '0';
+            int m = c - '0';
             a += m;
             b *= m;
         }
